requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
Django==3.2.7
django-rest-knox==4.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
joblib==1.0.1
numpy==1.21.2
Pillow==8.3.2
pycparser==2.20
pytz==2021.1
scikit-learn==1.0
scipy==1.7.1
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.0.0
whitenoise==5.3.0
